Update patch set 10
Patch Set 10:

Thanks for the review and suggestion Artem.

 > *I like the approach as such. However my personal preference would
 > be to define hypervisor (executor host) inside of a pool (a static
 > driver configuration-like approach). Otherwise for each executor
 > node I will need to add a new provider. For me it's more like I
 > have a servers pool (i.e. 5-10 nodes), which I would like to use
 > for zuul. Do I really need to define separate provider for each
 > node?
 > 

Yes, that sounds like a good idea.


 > * And then there is question how do I define base images for labels
 > (choose container image)? Or am I supposed to run a complete OS
 > image in a container (is it possible at all)?
 > 

Runc needs a rootfs (directory). By default the driver re-use the one (/)
from the host, though it is be possible to unpack a regular container
or diskimage into a sub directory.
Being able to manage non-openstack image with nodepool should be possible,
but it is not implemented yet, so it's up to the operator to manage
the available runC filesystems.


 > * Is it assumed, that each project need to also deliver
 > config.json?
 > 

Currently the project doesn't need to deliver any configuration,
the config.json is generated by the driver. Actually Nodepool
doesn't know which project requests a Node, it only acts on the
label name.
If you see use-cases for custom config.json, then we could make
it configurable through different label's name.


 > * Would you please add documentation with description and example
 > of use?

Yes sure, I was waiting for some feedback. I can add some in the next
PS.

Patch-set: 10
Label: Verified=0