Temporary fix static node assign race
Currently there is a race in the static node driver such it assigns
static nodes to more requests than allowed. This can be fixed
temporarily by using a launch lock on the provider.

Note that this only works if a provider is only existing on one
launcher. The real fix for this will be pre-registration of static
nodes.

Change-Id: Id5742b37dabf0c1badd0df70c5b5a521c95bdaac