Pre-register static nodes
This changes the static driver to register nodes from its config
file with ZooKeeper at startup and when the config file changes.
A node is registered max-parallel-jobs times with ZooKeeper (default
is 1).

Once a node is deleted, it is immediately re-registered.

Although this handles adding new nodes to the config, or increasing
the max-parallel-jobs count, it does not yet handle the reverse.

Change-Id: Ie386e440c2016af390087d109f7e5211c4fd968d