WIP: Fix for pools with diff labels
WIP because this needs a test. We can see the current test_multiple_pools
test get a KeyError during checks for image availability due to the bug,
but it still properly declines the request (but for a different reason).
That is not seen with this change (thus confirming the fix), but have to
figure out how to programmaticly verify that.

Pools within the same provider that have different labels are
attempting to handle requests for ALL labels defined within the provider.
This is due to change I61263f51be5e957de71d1e2dabaa7391bbe7bddf which
incorrectly changed from getting supported labels from per-pool to
per-provider. This modifies the getSupportedLabels() API to support
pools.

Change-Id: I7a0d472928c6b528f6faa6dd3b9cf1479874cb22