Update README to use python3 installation
It appears that nodepool does not work with python2,
so we update the README to instruct developers to use
python3 & pip3.

The bindep profile used for testing is also different,
so we amend it to correctly use the 'test' profile.

We also add intructions for using the fake provider
test scripts.

Change-Id: I6923a532e3a702def0595998ea2fab5c067da2fe