[POC] Add baremetal subimage upload support
This change adds rudimentary baremetal subimage upload support for user
images as described in:

https://docs.openstack.org/ironic/latest/install/configure-glance-images.html

In short, when building an image with a baremetal element, the kernel
and ramdisk is extracted from the main image, and uploaded as separate
sub-images to the provider. The main disk image's metadata then links
to the provider image ID's of the ramdisk and kernel.

This review is not currently in shape for merging, it's intended as a
rough POC.

Change-Id: Ic7dfdd9ddef7af02b5b7afe13f823ffb04dce28f
Signed-off-by: Jan Gutter <jan.gutter@netronome.com>
Acked-by: Frikkie Scholtz <frikkie.scholtz@netronome.com>