Allow launch-retries to indefinitely retry for openstack
When launch-retires for openstack provider is set to -1, nodepool will
indefinitely keep retrying to launch a node. This attempts to restore
nodepoolv2 functionality, where nodepool will conintue indefinitely to
bring a node online and never allow a job to hit a NODE_FAILURE
condition.

With a setting of 0, nodepool will not retry a launch and mark the
job as failed.

Change-Id: I50c085f692dd1646298b7ef146940e5bf500f5d5
Signed-off-by: Paul Belanger <pabelanger@redhat.com>