Update patch set 1
Patch Set 1:

(1 comment)

> (1 comment)
 > 
 > Thanks for starting this, I'm hoping to also do some baremetal
 > images with nodepool.
 
No probs. This is basically a sample to future generations intended as a dire warning.

 > I think what I would like to see here, is if a user needs to upload
 > 3 images for baremetal support (over single whole disk), is we
 > create 3 actually diskimages entries in nodepool.yaml and build
 > each image one at a time.

Aside from "momentum" (i.e. nobody's done 3 separate builds ever) there's also the problem of consistency. For something like packages, I think at least the kernel image would be the same, but I'm not so sure about the ramdisk.

I _think_ it won't be a problem if the on-disk ramdisk image and the uploaded glance image is out-of-sync, but I'm not quite sure. Will have to check.

It does mean that building a custom kernel in the element is going to be non-trivial.

Yet another hinky idea is to somehow make an explicit "extract" method? Pass an image and "extract" the kernel and ramdisk: i.e. onboarding the "baremetal" element explicitly into dib. _sigh_ so many options, none of them expressly appetizing.

 > We might have to group them together
 > some how, since 1 cannot work without another, but believe that
 > could be a configuration option for the diskimages stanza.
 
Yeah, one of the ideas I had was "nested" images. ARI and AKI images are definitely not standalone.

 > Then, we continue to use the meta data to dynamically include the
 > image ids for the images.

It's also possible to have a fixup-metadata step in the provider driver, triggering each time a baremetal element is uploaded, and fixing up the base image when the final element is uploaded, no matter in which order. Depending on where and how the image details is cached, this might just mean one extra call to the cloud.

Patch-set: 1