Update patch set 4
Patch Set 4: Code-Review-1

I think this is the point where we should split out the doc-requirements into doc/requirements.txt as it is done in openstack/nova.

This makes it possible to pull in zuul as a doc requirement without polluting the normal test environment. Further we won't have to install sphinx on every test run.

Despite this we might want to move the spninx stuff from zuul into zuul_sphinx.

Patch-set: 4
Label: Code-Review=-1