Update patch set 7
Patch Set 7: Code-Review-1

This patch also needs docs.

Also, it appears that we are relying on ~/.aws/credentials with [default] -- however, it is common to have multiple credentials in this file, i.e:

    [one]
    aws_access_key_id = one
    aws_secret_access_key = one

    [two]
    aws_access_key_id = two
    aws_secret_access_key = two

    [three]
    aws_access_key_id = three
    aws_secret_access_key = three

I somewhat expected the name of the provider (or the pool?) to match these -- I ended up using default for the time being but it could be improved.

Patch-set: 7
Label: Code-Review=-1