Use node cache in quota calculations
The quota calculations are racy by design and nodepool can safely deal
with that. Thus it doesn't matter if this information is outdated by a
few milliseconds so we can safely utilize the node cache during these
calculations.

Especially with many nodes in the system this speeds up the method
estimatedNodepoolQuotaUsed which is on the critical path in nodepool
request handling and currently limits scalability with many nodes and
few providers.

Change-Id: I97d289f3517220af3521a0b75717798382d9df09