Cache node iterations in cleanup workers
The DeletedNodeWorker does an async cleanup of nodes that are marked
for deletion. Therefor it can safely work on cached data to eliminate
the need of requesting every node in the system every five seconds.

The same can be done when cleaning up for the max hold age.

When resetting lost requests however this is not possible because we
need sequential consistency with the requested node request just
before the iteration. Make it clear that using the cache there is a
bad idea.

Change-Id: Ifb8702a411d5fe2a799477a51a2a0b6f5658d0a1