WIP: Try to fix single cloud config reload
With a single cloud a config reload can lead to failed requests
because the manager is stopped and replaced by a new one. In this
process any in-flight instance creations will fail and lead to
declines of this provider which is bad in the single cloud usecase.

To fix the single cloud usecase the idea is to catch the
ManagerStoppedException, abort the node and roll-back the changes to
the node request and not fulfilling or declining it. After unlocking
the node request it should be possible to retry it by the same but new
provider.

Change-Id: Ib1d036ac74ba1df738556cc173388cd38794e0a4