WIP: Second try to fix single cloud config reload
With a single cloud a config reload can lead to failed requests
because the manager is stopped and replaced by a new one. In this
process any in-flight instance creations will fail and lead to
declines of this provider which is bad in the single cloud usecase.

The idea to fix this is to check in poll if the manager is still
running. If it's not running anymore we can deallocate the nodeset and
unlock the node request.

Change-Id: Icd890a35775afa9aa0ac165c157cb9ead4fd7008