Use quota handling code for min-ready declines
We currently check the current pool nodes against max_servers in order
to decline min-ready request when we're at quota. This is older than
the quota handling where max-servers was the only limiting thing. Now
we have real quota handling we should defer this min-ready check to
the point where we know that we're at quota (independent of self
defined pool quota or real cloud quota).

Change-Id: Ic36ecfaf441436db6699297ef459c586f14c2dfc