Fix skipping of max-servers if not set
We currently check the max_servers for None and if it's not None we
validate the max-servers. However this gets initialized with math.inf
so this check is always true and we do an unneccessary max-servers
validation.

This validation can be an expensive operation because it iterates over
all pool nodes. Further this check is on the critical path that
assigns and more important polls and fulfills the node request
handler. In a system with many open node requests and many nodes and
few providers this can have a decent penalty on service times.

Change-Id: I8e8d4fd48267b34c66546e5030565dbd9b9c2b30