Update patch set 5
Patch Set 5:

>> How can we filter out functional test python file from the regular tox-py3 job?

Not sure I know what you mean. Are you suggesting we run the functional test in the tox-py3 jobs? We can't do that because it's a different tox target. I think we need separate targets (and jobs) for functional tests because each one is going to require separate setup (k8s install for k8s driver, etc).

Patch-set: 5