Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

I agree that this check fixes it such that we do not do a count of nodes when we know that the check "current_count >= self.pool.max_servers" will just fail anyway.

So if we want to keep the status quo, consider me +2 on this

However, per my inline comments, giving a -1 for discussion over what this is achieving.  I feel like we're encoding undocumented different behavior for clouds that we calculate the quota for v set max-servers for.  It doesn't seem to make much sense.

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1