Cleanup node requests that are declined by all current providers
In some corner cases (removing a provider) a node request can be left
in a state where all current providers did decline the request but one
that should have declined afterwards was removed from the
configuration before it had a chance to do so. Therefore we need to
check for requests that are declined by all currently running
providers and fail them if no provider is left.

Change-Id: I296cc28daaedd828520a366920b6b44d276ce487