Initialize label statistics to zero
The label statistics age gauges that keep their values. Currently
nodepool doesn't initialize the label based node statistics to
zero. Because of this any label state that shows up in the statistics
won't be reset to zero and nodepool is stopping to send updates to
this gauge. This leads to graphs that are mostly stuck to 1 if no
nodes are currently in this state.

This can be fixed by iterating over the supported labels and
initializing all states of them to zero.

Change-Id: I6c7f63f8f64a83b225386f6da567bfae5141be7b