Allow double-calling of TaskManager start
The most recent release of openstacksdk calls start on the TaskManager,
but we also call start on the TaskManager. In another few patches we'll
be getting rid of the nodepool TaskManager, but we're missing a patch in
openstacksdk to allow that. For now, just make it safe to call start
twice.

Change-Id: I893a6c03aa7cf9fa46a2a94daf3db3d4a7c9ddee