Update patch set 7
Patch Set 7:

Since the update of the cache is asynchronous, I think there may be a race in the test.  Maybe we could use iterate_timeout to wait 60 seconds for the values to be equal, otherwise fail?

I've looked around a bit, and I don't see any places where we are currently using nodeIterator where I would expect using the cache to cause a problem.

In fact, I think we might be able to be more aggressive and just change the default to cache=True.  I think perhaps the only times we really care that a node is completely up to date is when we have the lock.  So maybe we could put a direct getNode call inside the lockNode method, so that when we lock the node, we automatically synchronously update it, then just use cached data all the rest of the time?

Patch-set: 7
Label: Verified=0