Update patch set 5
Patch Set 5: Code-Review-1

(2 comments)

My feeling is the stat & graphing of this is enough for a -1, although it can be done as follow-on I'd not like to see us forget it.

My main concern was "is this a temporary weird thing we're seeing that we're putting into code to be maintained for eternity"; clarkb I think has convinced me it's not.  Maybe some comments would help anyone in the future wondering why we bothered.

Patch-set: 5
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1