Prepend exception output with time, date and thread
Currently, exceptions in the logs are put in "bare"; an example
probably serves best:

---
 2018-10-24 20:15:14,067 DEBUG nodepool.TaskManager: Manager ovh-gra1 ran task ComputeGetServers in 0.33750057220458984s
 2018-10-24 20:15:14,068 DEBUG nodepool.TaskManager: Manager ovh-gra1 running task ComputeDeleteServers (queue 3)
 2018-10-24 20:15:14,075 ERROR nodepool.NodeLauncher-0003396117: Request 200-0006852594: Launch attempt 1/3 failed for node 0003396117:
 Traceback (most recent call last):
   File "/usr/local/lib/python3.5/dist-packages/nodepool/driver/openstack/handler.py", line 227, in launch
     self._launchNode()
     ... several lines of blah blah ...
     :param block_device_mapping_v2: (optional) A dict of block
 openstack.cloud.exc.OpenStackCloudCreateException: Error creating server: 5dda5efe-6207-4433-a60d-90664f8a4844
2018-10-24 20:15:14,088 INFO nodepool.NodeLauncher-0003396117: Creating server with hostname  ...
---

This irregular format makes it a real pain to correlate useful
information; if you're investigating node 0003396117 and grep for
that, you don't get all the information on the failure you actually
need.  You need to go to the logs, search for "0003396117" and then
start following the lines manually.  In a situation like recently with
OVH, where you have multiple regions failing a large number of nodes,
it becomes quite annoying trying to correlate everything.

This custom formatter will mean that for a default log setup, we will
prefix all of the exception lines with the same info as the initial
message; e.g. above will become (lines trimmed, you get the idea)

---
 2018-10-24 20:15:14,067 DEBUG nodepool.TaskManager: Manager ovh-gra1 ran task ComputeGetServers in 0.33750057220458984s
 2018-10-24 20:15:14,068 DEBUG nodepool.TaskManager: Manager ovh-gra1 running task ComputeDeleteServers (queue 3)
 2018-10-24 20:15:14,075 ERROR nodepool.NodeLauncher-0003396117: Request 200-0006852594: Launch attempt 1/3 failed...
 2018-10-24 20:15:14,075 ERROR nodepool.NodeLauncher-0003396117    Traceback (most recent call last):
 2018-10-24 20:15:14,075 ERROR nodepool.NodeLauncher-0003396117     File "/usr/local/lib/python3.5/dist-packages/...
 2018-10-24 20:15:14,075 ERROR nodepool.NodeLauncher-0003396117       self._launchNode()
 2018-10-24 20:15:14,075 ERROR nodepool.NodeLauncher-0003396117       ... several lines of blah blah ...
 2018-10-24 20:15:14,075 ERROR nodepool.NodeLauncher-0003396117       :param block_device_mapping_v2: (optional)...
 2018-10-24 20:15:14,075 ERROR nodepool.NodeLauncher-0003396117     openstack.cloud.exc.OpenStackCloudCreateExcep...
 2018-10-24 20:15:14,088 INFO nodepool.NodeLauncher-0003396117: Creating server with hostname  ...
---

Note it has all the same prefix as the message, as it's from the same
record.  So grepping by node id, timestamp or even 'ERROR' level logs
is going to contain all the useful info you need.

Change-Id: I9d8946b9b963d5ed3e93fb141a3f8d7044c48e3f