Rate limit updateNodeStats
We currently update the node statistics on every node launch and
delete. On busy systems this can be quite expensive and reduce the
throughput of node requests. This can be limited by updating the node
stats at max every 15 seconde which should be precise enough.

In a local test case fulfilling 200 node requests this reduced the
test execution time from 80s to 50s on my laptop.

Change-Id: I27f3ce42da6972640be79462d9906252fea78e93