WIP: Add request handler timer to stats
When a nodepool deployment grows with several providers and static
providers there will be a point in time when it makes sense to run
multiple launchers and shard the providers among them. One key metric
for this decision is the runtime of the request handler. With many
provider/pools and many nodes and requests in the system this leads to
a high thread contention and thus potentially very slow processing of
the request handler. This can lead to the situation that even if we
are not at quota we aren't able to create nodes quickly enough to fill
the quota.

The reason for such slowness is currently hard to debug and needs
in-depth log analysis to find out that the request handling is
slow. This adds a timer metric with the runtimes of the request
handler per provider. This can be valuable to judge from monitoring if
an additional nodepool instance is needed and providers must be
splitted.

Change-Id: I145977ec7a8ef9828f07f7ecc7c39356c29f0598