Add resource metadata to nodes
We currently don't really have statistics about the resource
utilization of projects or tenants in zuul. This can be solved by
adding resource metadata to the node data structure that can be used
by zuul (which has the knowledge about projects and tenants) to report
resource usage statistics.

The new introduced resources field is a plain dict and could be
extended in the future with further resource information like ports or
volume size that might be useful in the future.

Change-Id: I670d4bfd7d35ce1109b3ee9b3342fb45ee283a79