Update patch set 15
Patch Set 15:

Updated a comment regarding caching of node stats.

Regarding the update on lock idea. First sorry that I didn't respond to this. I think we should do that but in its own change. The reason for this is that caching doesn't change the guarantees  around locking but just increases the probability that the node is out of date. But centralizing the node update after lock absolutely makes sense in both reducing complexity and reducing sources of mistakes while using locking. I'll see if I can change that in a follow up.

Patch-set: 15