Update patch set 3
Patch Set 3:

Unsure if this is "ready", but generally LGTM. One thing I'd like for us to remember on changes to the ZK schema is compatibility with existing schema. We almost always have issues upgrading nodepool on schema changes. I don't *see* anything here that would require a total shutdown of nodepool to upgrade to this version, but would like others to keep this in mind as they review (perhaps you'll see something I did not)

Patch-set: 3