[wip] Add Fedora 29 testing
Replace Fedora 28 with Fedora 29 functional testing.

Note this changes our Red Hat platforms to use NetworkManager for
interface configuration, rather than legacy scripts.  Fedora 29 has
split the legacy scripts into a new package and it marked for future
removal.  NetworkManager is the default on Centos 7 and will also be
on Centos 8, so it makes sense to use it there too.

Depends-On: https://review.openstack.org/619120
Change-Id: I640838c68a05f3b22683c1e90279725a77678526