Cache node request zNodes
When implementing dynamic priority queues in nodepool we need to cache
the requests. This introduces a simple node request caching using the
TreeCache supplied by kazoo. This currently only caches the raw
zookeeper data and still requires json parsing when traversing the
requests. In a later change we can introduce a second layer of caching
that updates the NodeRequest objects in place so we can avoid
excessive json parsing.

Change-Id: I84b3474636ea1f14faa6bf7c8a6c6f83598c3741