Add second level cache of nodes
An earlier change introduced zNode caching of Nodes. This first
version still required frequent re-parsing of the node json data. This
change introduces a second level cache that updates a dict of the
current nodes based on the events emitted by the TreeCache. Using this
we can reduce json parsing and make node caching mode effective.

Change-Id: I4834a7ea722cf2ac7df79455ce077832ae966e63