Update patch set 2
Patch Set 2:

We should likely clarify this a bit more in the docs, but...

diskimages.meta - I think it's pretty clear this is for diskimages and not nodes

labels.instance-properties - This is an OpenStack driver specific set of metadata that will appear on the instance itself (in the data stored by nova). This does not appear in ZooKeeper.

The proposed new metadata is designed to appear only in ZooKeeper on the Node znode, and is thus useful for *any* backend driver in use by nodepool.

Patch-set: 2