Update patch set 2
Patch Set 2:

> We should likely clarify this a bit more in the docs, but...
 > 
 > diskimages.meta - I think it's pretty clear this is for diskimages
 > and not nodes
 > 
 > labels.instance-properties - This is an OpenStack driver specific
 > set of metadata that will appear on the instance itself (in the
 > data stored by nova). This does not appear in ZooKeeper.

I don't think this needs to be openstack specific. docker and k8s "instances" also have the idea of metadata. We likely want to expose this metadata to those drivers at they are added as well. Note that the consumption of the metadata differs across providers but there is some mechanism.

The other thing I would suggest is that we use the same rough datastructure for the different metadata values. Currently I think it is:

  metadata:
    key: value
    key': value'
    key'': value''

And does not have nested lists.

 > 
 > The proposed new metadata is designed to appear only in ZooKeeper
 > on the Node znode, and is thus useful for *any* backend driver in
 > use by nodepool.

Patch-set: 2