Block 0.19.0 of openstacksdk
There is a bug in Rackspace Public Cloud that causes keystoneauth to
fail doing discovery which 0.19.0 of openstacksdk exposes because it
starts using keystoneauth discovery directly. Until the keystoneauth fix
lands and is released, running nodepool with 0.19.0 of openstack will
fail when attempting to use Rackspace Public Cloud.

Just for the record, this is due to the fact that Rackspace Public
Cloud:

- has invalid integer project ids
- still senselessly keeps them in the compute service URL
- blocks access to the compute discovery document

The keystone team are kindly accepting a workaround fix to keystoneauth
even though it is a workaround for what is a completely invalid setup.

Change-Id: I72ec16ecb7770d97aa5703bdcfd3e8b188c89f19