Fix race when deleting Node znodes
There is a race where:

  Thread A               Thread B
  locks the node
  deletes the instance
  deletes the lock
                         locks the node
  deletes the node data
                         loads empty data from the node

Thread A only deletes the node data because during it's recursive
delete, a new child node (the lock from B) has appeared.  Thread B
proceeds with invalid data and errors out.

We can detect this condition because the node has no data.  It may
be theoretically possible to lock the node and load the data before
the node data are deleted as well, so to protect against this case,
we set a new node state, "deleted" before we start deleting anything.

If thread B encounters either of those two conditions (no data, or
a "deleted" state), we know we've hit this race and can safely attempt
to recursively delete the node again.

Change-Id: Iea5558f9eb471cf1096120b06c098f8f41ab59d9