Fix race in test_handler_poll_session_expired
Just waiting for a node request state change does not guarantee
that we have actually entered the poll() portion of request handling.
This was causing the mock.call_count assert to fail on occasion.
Change this to simply wait on the call_count to increment, eliminating
that race.

There was also another potential race in checking that the request
handler was removed from the request_handlers structure. It would
be possible for the same request to re-enter active handling before
we had a chance to check that it was removed when the session
exception was thrown. We handle that race by setting the request state
to FAILED on the first exception.

Change-Id: I646b82243eb7f8c4e83d6678c2c0d265d99e51e0