Switch devstack jobs to Xenial
Change I8749ed24d5f451d29f767ebb2761abd743b7d306 modified the devstack
based jobs to run on Bionic.  Unfortunately we're not quite ready for
that; one issue is that our devstack dependencies require zypper for
opensuse which is not on bionic.  diskimage-builder excludes zypper on
bionic in bindep, but we don't have a mechanism to use that (yet [2]).

For now, switch them back to Xenial to retain the status quo.  We can
then take a more controlled approach to work on modernising them.

[1] https://bugs.launchpad.net/ubuntu/+source/zypper/+bug/1808230
[2] https://review.openstack.org/624852

Change-Id: I56646e49264dd844f5818a84e04965863542f572