Update patch set 35
Patch Set 35:

> > Wouldn't it be useful to have nodepool log files to help with
 > > debugging issues when they pop up?
 > 
 > Yes it would be useful, but this doesn't work anymore since
 > nodepool is now started by tox. Where are the logs written now?
 > Wouldn't it be better if functional jobs use a common parent job to
 > retrieve the logs?

Oops, yeah, we already have nodepool logs in the tox output. I think I was mostly considering the other service logs (OpenShift in particular) and stupidly lumped nodepool in with my thinking. Are the OpenShift logs useful to grab? If not, then we don't need the post playbook.

Patch-set: 35
Label: Verified=0