Add a timeout for the image build
A builder thread can wedge if the build process wedges. Add a timeout
to the subprocess. Since it was the call to readline() that would block,
we change logging to the build log file slightly to have the subprocess
write directly to that in buffered chunks. This allows us to set a
timeout in the Popen.wait() call.

The timeout value can be controlled via a command line option to the
builder process and defaults to 8 hours.

Change-Id: I188e8a74dc39b55a4b50ade5c1a96832fea76a7d