Update patch set 3
Patch Set 3:

Hi John,

This looks like a cool driver idea! Can't wait to see how it turns out.

You seem to be making fixes for pep8 errors in new reviews. That isn't really a valid workflow since we'll never be able to approve and merge this original review that way. You'll need to correct the errors in this review and resubmit as a new patchset. You can read a bit more about that in the "Updating a Change" section in:

https://docs.openstack.org/infra/manual/developers.html#development-workflow

Visit us in the Freenode IRC channel #zuul if you need assistance with this workflow.

Patch-set: 3
Label: Verified=0