Update patch set 12
Patch Set 12:

> > I tried p.kill() and p.terminate() here. Both cause the builder
 > to hang, and I'm not entirely certain why. Perhaps because of the
 > signal handling we are already doing? If you know how to solve
 > that, I'd be happy to add it here.
 > 
 > Hrm, i was actually thinking SIGINT but SIGTERM should,
 > theoretically work.  When you say the builder hangs, you mean the
 > nodepool builder process, not the diskimage-builder process?
 > 
 > I just feel like if we didn't even try to exit things cleanly in
 > dib, the system is almost guaranteed to be in a system-wide crappy
 > state with orphaned temporary files (and eventual full disk) and
 > abandoned mounts.

Yes, nodepool-builder hangs.

I don't think we need to solve that issue in this change, since we already have that problem anyway, and it is likely going to call for a much more in-depth change to handle properly. This review at least improves on the current situation a bit.

Patch-set: 12