Stop building an explicit nodepool-base image
We have a utility image that we use for running the nodepool command
that doens't have any additional software installed. Although it does
set a COMMAND of /usr/local/bin/nodepool, it could still be useful as
a general base image for other people if they wanted such a thing.

Change-Id: I894e3d2dbe3cd2017f27ccc5e6fe298e9c9abd03