openstack: Implement user data configuration
When using cloud images, rather than disk images, it is often
useful to make use of cloud-init to prepare the node in some
way for use. This could involve adding software, or changing
configuration.

To make this possible, we add a new configuration option called
user_data which is only used for cloud images.

Change-Id: Ibf26752ab21c35babd9ad29f92fd95080d9e9a52