Add openstackci-mirrors element for centos/ubuntu functional tests
This dib element configures the image builds to use local mirrors in
the Open Stack gate.  This should improve the reliability of this
testing by eliminating random network failures.

The dib-setup-gate-mirrors role is used to create the bootstrap repos
(that point to the local mirror).  dib needs to be pointed to these
repos; since dib is called by nodepool we need to update the nodepool
config file; since the nodepool config file is created by the devstack
plugin we need to pass this through local.conf.

Depends-On: https://review.openstack.org/634621
Change-Id: I9b994e65c762fe2296d2d07bc8ab87b32e6f4b9f