Update patch set 17
Patch Set 17:

(6 comments)

Thanks for the feedback Jan. Zuul console stream actually needs a patch in zuul to support a custom path for the zuul_stream callback. The current design is that:
A single zuul_console daemon is started on the "hypervisor", streaming file from /var/tmp/zuul-console/ , and each containers starts with the shared directory bind mounted. Thus to make the stream work, you need: https://review.openstack.org/535538 and set log_stream_file=/var/tmp/zuul-console/{log_uuid}.log in the [executor] section.

I was hoping the zuul streaming would be rewritten by now (using ssh port forward to remove the need of that zuul_console/zuul_stream), but perhaps it would be faster to just fix the comment on 535538 and move on with this solution. The issue with zuul_console/stream is that it relies on a long lived daemon listening on a static port, thus when multiple jobs run on the same ip (but in different container), we need a way to share the console log file with the daemon. 

Alternatively, if you don't mind sharing the host /tmp with all the container, you could use the regular /tmp/{log_uuid}.log path by replacing the /var/tmp/zuul-console bind mount (L344 of the provider.py module) with /tmp and removing the zuul_console path argument L62 of the init.yml file. In that case, no need to patch or configure zuul.

Patch-set: 17