Update patch set 2
Patch Set 2:

Outer labels are not readily available in the handler context, and most label settings are already defined in provider, e.g. username. Thus I think it makes more sense to keep the logic in provider.

PS3 refactor the code to reduce the amount of duplication.

Patch-set: 2
Label: Verified=0