runc: Allow overlayfs mounts for container's rootfs
The old settings just used a read-only root, with a bind-mount for a
writable user's home. This is not enough for, say, installing additional
packages via job-specific playbooks.

There's some support for overlayfs directly within `runc` [1], but it
isn't obvious to me how to enable that feature for rootfs. In the OCI
specification, the `root` option is required and it is supposed to point
to an already existing directory on host. There are no mount options for
rootfs directly. The `mount` stanzas control mounts underneath the
container's rootfs, something which is not suitable here.

Given that we're already using Ansible to prepare bits on hypervisor,
let's use it to prepare the overlay mount as well.

[1] https://github.com/opencontainers/runc/pull/1314/commits/ca14e7b463101d4b13f7c23dd952ffe056e6a22c

Change-Id: I178306f10a7c3449ba49eb543fcc6e9531a401fc