Update patch set 4
Patch Set 4: Code-Review-1

There is a hierarchy image > provider > pool, and we should put attributes as far "left" in that as possible to reduce duplication.

This is, ultimately, an attribute of the image being used.  It doesn't change from one pool to the next.  It's very much like the 'username' attribute, and I think a good rule of thumb is that we should probably put it wherever we put 'username'.

So in the case of the VM providers, it should be at attribute of the diskimage or the cloud-image.

In the static provider, it should be an attribute of the node.

In the container providers, it will need to stay on the label as it is in this change, which, I agree, is awkward.

Also, we should probably add this to some test config files.  :)

Patch-set: 4
Label: Code-Review=-1