Support requests for specific providers
In Zuul we support a paradigm where a job can be paused, whereupon
dependent jobs then request nodes and are started.  It is nearly
always the case that the user would like the nodes for the dependent
jobs to be in the same provider as the parent, as the use cases
generally involve transferring data between the two.

To support this, add a 'provider' attribute to the node request which,
if present, means that all on-line launchers for that provider must
decline the request before anyone else processes it.  This will cuase
the desired behavior if everything is working, and if some calamity
befalls that launcher, other launchers can still attempt to fulfill
the request, which might work, or might not, but performing that last
ditch effort is fine once there are no alternatives.

Change-Id: I91fe05081695d454651f6068eac5c08ac30ff899