Fix loss of ZK conn during node delete
NOTE: This currently demonstrates the test failure until we fix it.

This adds a test to validate that if we lose our ZK session (or
otherwise throw any exception from deleteRawNode()) while deleting
a node, and AFTER we've already set the state to DELETED, that we
successfully remove the node left in the DELETED state later on in
the DeletedNodeWorker thread.

Change-Id: Ie6c00e1adcbdf5b6f2122fc638ee5e91aaca9f61