Update patch set 1
Patch Set 1: Code-Review-2

Hi, thanks for the change, but at the moment we are not ready to support out-of-tree drivers in Nodepool (though we may at some point in the future).  I'm not sure the API is stable enough yet, and we'd like Nodepool to work with lots of providers out of the box, so we're generally in favor of adding providers which are generally useful and have tests.

What sort of driver are you looking to support?

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2