Update patch set 1
Patch Set 1: Code-Review-1

A careful reading of the code indicates that the existing code *does* do something.  We can certainly consider changing this, but this commit (the comments and commit message) is apparently written with the assumption that the code being bypassed is dead.  That's not the case, and it should not be reviewed in that context.

The current code verifies that, at the least, nodepool can make a connection to the remote system (which verifies that it is booted, the network subsystem is up, etc).

If you want to change that, feel free to propose that, but please properly contextualize that change.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1