Update patch set 1
Patch Set 1: Code-Review-1

LGTM, but it will break us because we're currently relying on calling diskimage-create from the PATH so we can switch out diskimage-create per diskimage (by setting the PATH variable) by wrapper scripts that do either preprocessing of the elements (updating custom elements on demand) or calling other tools for creating the images (e.g. windows).

I know that what we're doing is kind of a hack. A solution for this could be to let the user choose the diskimage-create script in the diskimages. What do you think about this? At least it would be great if we could find an upstreamable solution for this workflow before we land this change.

Patch-set: 1