wip: Allow proceeding with requests on quota exceeded
Due to it's nature the static driver can not spawn additional nodes as
e.g. the OpenStack driver. For this reason the static handler will
always tell that is has no remaining quota.

The current implementation will pause all further request handling,
since it assumes that the quota limit applies to all node types of a
pool. This however wrong in the case of the static provider. Just
because one node type is not available doesn't mean that other node
types are also not available.

With the proposed change the request can allocated needed types from the
currently available nodes and then go into a 'waiting' state until the
missing node types become available.

Change-Id: If69cbeb33b8abb4f8e36146b9d849403dccb7d32