Update patch set 2
Patch Set 2: Code-Review-1

This isn't necessary to test nodepool functionality, so we shouldn't be writing this kind of change for this repo.  How about you add an option to the job to pass in more of the image building config (including the list of elements).  I think that's reasonable to merge here, then you won't need to change this job to test new elements in the future.

(Also, just for my own curiousity, I thought there were other ways to test individual elements?)

Patch-set: 2