Update patch set 2
Patch Set 2:

> This isn't necessary to test nodepool functionality, so we
 > shouldn't be writing this kind of change for this repo.  How about
 > you add an option to the job to pass in more of the image building
 > config (including the list of elements).  I think that's reasonable
 > to merge here, then you won't need to change this job to test new
 > elements in the future.
 > 
 > (Also, just for my own curiousity, I thought there were other ways
 > to test individual elements?)

So I think the commit message is a bit misleading here. The goal isn't to test the journal to console element as much as it is to use the journal to console element to understand why nodepool + glean + dib + centos is so broken.

I think this is useful debugging information even if nodepool doesn't directly depend on it for when the jobs do fail. Is there a problem with adding debug info to the nodepool job by default for when the jobs do fail?

Patch-set: 2
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0