Update patch set 3
Patch Set 3:

Yes, when there are more static providers that supply the same labels one could hold many waiting request while the others sit idle. This might need to be change, but should at least be documented.

Regarding a separate pool for each label: that's exactly what we are currently using and there are some problems with this approach.

First of all we need to be able to assign multiple labels per static node. With this requirement it's not possible to partition the nodes by label (maybe we could have a pool for each label-set, but then you will run into the problem below).

Then there is also the problem when then last node in a pool is removed or relabeled. This node will still exist in Zookeeper but there won't be a pool anymore to take care of the cleanup. Currently those orphans need to be removed manually before a node can be reused with a different label.

Maybe some background on how we use static nodes: most of those nodes are test equipment connected to some device under test (DUT). Those DUTs have certain capabilities that are expressed as labels. The DUTs also need to be maintained or replaced from time to time. That's why there is a need for label changes.

Not pausing the pool was just a first stab at solving our immediate problems and to start the discussion. I'm completely open to solving this another way.

Patch-set: 3